+++ src/identity_mapper/requests.py
@@ -2,6 +2,7 @@ from identity_mapper.capability_protocol import (
     AuthenticateRequest,
     MapIdentityRequest,
     ResolveIdentityRequest,
+    ResolveTargetIdentityRequest,
     VerifyCredentialRequest,
 )
 
@@ -9,5 +10,6 @@ __all__ = [
     "AuthenticateRequest",
     "MapIdentityRequest",
     "ResolveIdentityRequest",
+    "ResolveTargetIdentityRequest",
     "VerifyCredentialRequest",
 ]