+++ src/identity_mapper/responses.py
@@ -2,6 +2,7 @@ from identity_mapper.capability_protocol import (
     AuthenticateResponse,
     MapIdentityResponse,
     ResolveIdentityResponse,
+    ResolveTargetIdentityResponse,
     VerifyCredentialResponse,
 )
 
@@ -9,5 +10,6 @@ __all__ = [
     "AuthenticateResponse",
     "MapIdentityResponse",
     "ResolveIdentityResponse",
+    "ResolveTargetIdentityResponse",
     "VerifyCredentialResponse",
 ]